+++ wagtailsharing/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 
+from wagtail.contrib.routable_page.models import RoutablePageMixin
 from wagtail.core.models import Site
 
 
@@ -44,3 +45,16 @@ class SharingSite(models.Model):
             return "https://{}".format(self.hostname)
         else:
             return "http://{}:{:d}".format(self.hostname, self.port)
+
+
+class ShareableRoutablePageMixin(RoutablePageMixin):
+    def route(self, request, path_components):
+        if getattr(request, "routed_by_wagtail_sharing", False):
+            page = self.get_latest_revision_as_page()
+            # This call to RoutablePageMixin's route() is so that the  method
+            # gets called with the latest-revision-as-page object as self,
+            # rather than the page object that is current self in this context.
+            # This ensures that, if we're being routed by wagtail sharing, we
+            # serve the latest revision.
+            return RoutablePageMixin.route(page, request, path_components)
+        return super().route(request, path_components)