+++ wagtailsharing/tests/test_models.py
@@ -4,6 +4,10 @@ from django.test import RequestFactory, TestCase
 from wagtail.core.models import Site
 
 from wagtailsharing.models import SharingSite
+from wagtailsharing.tests.shareable_routable_testapp.models import (
+    RoutableTestPage,
+    ShareableRoutableTestPage,
+)
 
 
 class TestSharingSite(TestCase):
@@ -84,3 +88,54 @@ class TestSharingSite(TestCase):
     def test_root_url_other_port_http(self):
         site = SharingSite(hostname="test.hostname", port=1234)
         self.assertEqual(site.root_url, "http://test.hostname:1234")
+
+
+class TestShareableRoutablePage(TestCase):
+    def setUp(self):
+        self.factory = RequestFactory()
+        self.default_site = Site.objects.get(is_default_site=True)
+        self.sharing_site = SharingSite.objects.create(
+            site=self.default_site,
+            hostname="sharinghostname",
+            port=1234,
+        )
+
+        self.root_page = self.default_site.root_page
+        self.routable_page = RoutableTestPage(
+            title="Routable page", text="Published text", live=True
+        )
+
+        self.shareable_routable_page = ShareableRoutableTestPage(
+            title="Shareable routable page", text="Published text", live=True
+        )
+
+        self.root_page.add_child(instance=self.routable_page)
+        self.routable_page.text = "Draft text"
+        self.draft_revision = self.routable_page.save_revision()
+
+        self.root_page.add_child(instance=self.shareable_routable_page)
+        self.shareable_routable_page.text = "Shareable draft text"
+        self.draft_revision = self.shareable_routable_page.save_revision()
+
+    def test_route_not_sharing(self):
+        plain_response = self.client.get("/routable-page/")
+        self.assertContains(plain_response, "Published text")
+
+        shareable_response = self.client.get("/shareable-routable-page/")
+        self.assertContains(shareable_response, "Published text")
+
+    def test_route_with_sharing(self):
+        # Request from the sharing site
+        plain_response = self.client.get(
+            "/routable-page/",
+            HTTP_HOST="sharinghostname",
+            SERVER_PORT=1234,
+        )
+        self.assertContains(plain_response, "Published text")
+
+        shareable_response = self.client.get(
+            "/shareable-routable-page/",
+            HTTP_HOST="sharinghostname",
+            SERVER_PORT=1234,
+        )
+        self.assertContains(shareable_response, "Shareable draft text")